fix: timeselectinput component
