+++ src/components/TimeSelectInput/ampmSelect.js
@@ -1,10 +1,8 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import { uniqueId } from './../../libs/utils';
-import isChecked from './helpers/isChecked';
-import StyledInputHidden from './styled/inputHidden';
-import StyledOptionLabel from './styled/optionLabel';
-import StyledSelectValue from './styled/selectValue';
+import { isChecked } from './helpers';
+import { StyledInputHidden, StyledOptionLabel, StyledSelectValue } from './styled';
 
 function handleAmPmBlur(event) {
     event.stopPropagation();
