+++ src/components/TimeSelectInput/icons/index.js
@@ -1,0 +1,4 @@
+import DownArrow from './downArrow';
+import UpArrow from './upArrow';
+
+export { DownArrow, UpArrow };
