--- src/components/TimeSelectInput/styled/container.js
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-const StyledContainer = styled.div`
-    width: 100%;
-`;
-
-export default StyledContainer;
