--- src/components/TimeSelectInput/styled/downArrow.js
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-import DownArrow from '../icons/downArrow';
-import attachThemeAttrs from '../../../styles/helpers/attachThemeAttrs';
-
-const StyledDownArrow = attachThemeAttrs(styled(DownArrow))`
-    color: ${props => props.palette.brand.main};
-`;
-
-export default StyledDownArrow;
